+++ json_serializable/lib/src/json_part_builder.dart
@@ -2,9 +2,6 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-// https://github.com/dart-lang/sdk/issues/31761
-// ignore_for_file: comment_references
-
 import 'package:build/build.dart';
 import 'package:source_gen/source_gen.dart';
 
@@ -22,21 +19,6 @@ import 'json_serializable_generator.dart';
 /// [Map] and [List] instances created during serialization. This will
 /// increase the code size, but it may improve runtime performance, especially
 /// for large object graphs.
-///
-/// Usage:
-///
-/// ```dart
-/// new BuildAction(
-///   jsonPartBuilder(),
-///   'my_package',
-///   inputs: const ['lib/*.dart']
-/// )
-/// ```
-///
-/// See the [example] to understand how to configure your project for
-/// `json_serializable`.
-///
-/// [example]: https://github.com/dart-lang/json_serializable/tree/master/example
 Builder jsonPartBuilder({String header, bool useWrappers: false}) {
   useWrappers ??= false;
   return new PartBuilder([